#35 Testing what happens when queen is pressed
I was able to create an unique ID for the black queen and have it do
something. It changes the black king into a white tile. If every piece
is given an unique ID then the problem of multiple selections of the
same piece does not exist anymore, however more testing is needed.