#40 Highlights Rook's moves
Currently when a rook is clicked on, it will show anywhere the rook can
move, if its an empty space, the border of the tile is highlighted blue,
if it's an enemy piece it will be red. The highlighted pieces do not
account for collision and can be highlighted even if its a spot behind a
piece.