#46 Got the pieces to move to highlighted spaces
When a piece is clicked, highlighted tiles shows where the rook can move
to (#40). Now when a highlighted tile is clicked, the rook will move
there. Currently, it will allow the rook to look past it's ally pieces
for potential moves except when moving to the right.

Moving but technically it still does not work with collision with ally
pieces, when I applied the same methods for highlighting right moves to
left, the for loop is stuck on 0 and I cannot figure out why it doesn't
run through the if statements. This makes it so that highlighting the
other directions (that occur after the current loop (highlighting left))
never run. If the same method for moving right is applied to moving
upwards, then moving down, left, and right will not be displayed since
it is stuck on the for loop that is making the highlights for moving
forwards