User Story #85 Task 2
Currently 3 special events spawn at a time per round (after the black
player goes). This is for testing purposes and will be changed later
when special events are complete. Special event cannot be on top of a
existing piece nor move. It will always spawn on a random tile that is
l.png or d.png.