User Story #90 Transform Piece
Transform event now works. When a piece lands on the event, that piece
randomly changes into a random piece that is not a king. Currently the
back end changed but the front end does not. Meaning visually the board
does not update until the end of the next player's turn (based on our
update method).